+++ cypress/integration/transfer-error-handling.spec.js
@@ -20,4 +20,27 @@ describe('Transfer Error Handling', () => {
             .contains('Error')
             .log('Alert-box message informs user that an error occurred while fetching the metadata of the mappings')
     })
+
+    it('Should show an alert when an error occurred when logging out the data providers', () => {
+        // Intercept POST calls to /logout endpoint
+        cy.intercept('/logout', {
+            statusCode: 500,
+            status: 500
+        }).as('postCallLogout')
+
+        cy.visit('/')
+
+        // Find & expand the Settings card
+        cy.get('[data-test=card-header-settings]')
+            .click()
+            .log('Expanded Settings card')
+
+        // Click the button for logging out and assert that the POST call has been made
+        cy.get('[data-test=button-logout]')
+            .click()
+            .wait('@postCallLogout')
+
+        cy.get('[data-test=alert-box]').should('contain.text', 'Error')
+
+    })
 })